refactor: KEEP-1587 extract shared DB helpers into keeperhub-executor/lib/
Move duplicated functions from in-process.ts and workflow-runner.ts
into lib/db-helpers.ts as a single source of truth:
- updateExecutionStatus (with toJsonSafe serialization)
- initializeExecutionProgress
- computeNextRunTime
- updateScheduleStatus
- DbSchema type

Both in-process.ts and workflow-runner.ts now import from the shared
module instead of maintaining their own copies.
